rewrite lots more commands to ts
