--- src/commands/OSRS_Accounts/hcim.js
@@ -1,23 +1,0 @@
-const { Command } = require('klasa');
-const { Hiscores } = require('oldschooljs');
-
-module.exports = class extends Command {
-	constructor(...args) {
-		super(...args, {
-			cooldown: 2,
-			aliases: ['hc'],
-			description: 'Shows the stats of a HCIM account.',
-			usage: '(username:rsn)',
-			requiredPermissions: ['EMBED_LINKS']
-		});
-	}
-	async run(msg, [username]) {
-		const player = await Hiscores.fetch(username, { type: 'hardcore' }).catch(err => {
-			throw err.message;
-		});
-
-		const embed = this.getStatsEmbed(username, 11995146, player);
-
-		return msg.send({ embed });
-	}
-};
