+++ src/util.js
@@ -1,33 +1,3 @@
-const convertLVLtoXP = lvl => {
-	let points = 0;
-
-	for (let i = 1; i < lvl; i++) {
-		points += Math.floor(i + 300 * Math.pow(2, i / 7));
-	}
-
-	return Math.floor(points / 4);
-};
-
-const convertXPtoLVL = (xp, cap = 99) => {
-	let points = 0;
-
-	for (let lvl = 1; lvl <= cap; lvl++) {
-		points += Math.floor(lvl + 300 * Math.pow(2, lvl / 7));
-
-		if (Math.floor(points / 4) >= xp + 1) {
-			return lvl;
-		}
-	}
-
-	return cap;
-};
-
-const xpLeft = xp => {
-	const level = convertXPtoLVL(xp);
-	if (level === 99) return 0;
-	return (convertLVLtoXP(level + 1) - xp).toLocaleString();
-};
-
 function roll(max) {
 	return Math.floor(Math.random() * max + 1) === 1;
 }
@@ -66,9 +36,6 @@ const formatDuration = ms => {
 };
 
 module.exports = {
-	convertLVLtoXP,
-	xpLeft,
-	convertXPtoLVL,
 	roll,
 	cleanString,
 	fmNum,
